apple seeds fix

+++ code/game/turfs/floor_attackby.dm
@@ -364,7 +364,10 @@
 			else if (istype(C, /obj/item/stack/farming/seeds/apple))
 				visible_message("[user] places the seeds in the ploughed field.")
 				new/obj/structure/farming/plant/apple(src)
-				qdel(C)
+				if (C.amount>1)
+					C.amount -= 1
+				else
+					qdel(C)
 				return
 			else if (istype(C, /obj/item/stack/farming/seeds/orange))
 				visible_message("[user] places the seeds in the ploughed field.")
